wip: browser manager with firefox
